requirements.txt
argon2-cffi==16.2.0
asn1crypto==0.24.0
bcrypt==3.1.7
beautifulsoup4==4.6.3
certifi==2019.9.11
cffi==1.14
contextlib2==0.5.5
cryptography==2.7
Django>=2.0.0
django-extensions==1.9.9
django-passwords==0.3.12
django-simple-aes-field==0.1.3
django-webpack-loader==0.6.0
djangorestframework
djangorestframework-bulk
fabric==2.5.0
html5lib==1.0.1
invoke==1.3.0
ipaddress==1.0.22
oauthlib
paramiko==2.6.0
psycopg2
pycparser==2.17
pycrypto==2.6.1
PyNaCl==1.3.0
pytz==2019.1
raven==6.9.0
requests
requests-oauthlib
sentry-sdk
SimpleAES
six==1.10.0
South==1.0.2
typing==3.5.2.2
urllib3
webencodings==0.5.1
